### 2020-07-22 #2502 @whedon
[PRE REVIEW]: 21cmFAST v3: A Python-integrated C code for generating 3D realizations of the cosmic 21cm signal.
**Submitting author:** @steven-murray (<a href="http://orcid.org/0000-0003-3059-3823">Steven Murray</a>)
**Repository:** <a href="https://github.com/21cmFAST/21cmFAST" target ="_blank">https://github.com/21cmFAST/21cmFAST</a>
**Version:** v3.0.0
