### 2020-07-23 12430ed
Add stand-alone function to calculate FFTW wisdoms (minimises lines of code)
