+++ src/py21cmfast/src/GenerateICs.c
@@ -282,7 +282,7 @@ int ComputeInitialConditions(
     // FFT back to real space
     if(user_params->USE_FFTW_WISDOM) {
         // Check to see if the wisdom exists, create it if it doesn't
-        sprintf(wisdom_filename,"complex_to_real_DIM%d_NTRHEADS%d.fftwf_wisdom",user_params->DIM,user_params->N_THREADS);
+        sprintf(wisdom_filename,"complex_to_real_DIM%d_NTHREADS%d.fftwf_wisdom",user_params->DIM,user_params->N_THREADS);
         if(fftwf_import_wisdom_from_filename(wisdom_filename)!=0) {
             plan = fftwf_plan_dft_c2r_3d(user_params->DIM, user_params->DIM, user_params->DIM,
                                          (fftwf_complex *)HIRES_box, (float *)HIRES_box, FFTW_WISDOM_ONLY);
@@ -1019,6 +1019,111 @@ int ComputeInitialConditions(
         gsl_rng_free (r[i]);
     }
     LOG_DEBUG("Cleaned Up.");
+    } // End of Try{}
+
+    Catch(status){
+        return(status);
+    }
+    return(0);
+}
+
+int CreateFFTW_Wisdoms(struct UserParams *user_params, struct CosmoParams *cosmo_params) {
+
+    int status;
+
+    Try{ // This Try wraps the entire function so we don't indent.
+
+        // Put this check here to minimise checks on Python side
+        if(user_params->FFTW_Wisdom) {
+            Broadcast_struct_global_UF(user_params,cosmo_params);
+
+            fftwf_plan plan;
+
+            char wisdom_filename[500];
+
+            int i,j,k;
+
+            omp_set_num_threads(user_params->N_THREADS);
+            fftwf_init_threads();
+            fftwf_plan_with_nthreads(user_params->N_THREADS);
+            fftwf_cleanup_threads();
+
+            // allocate array for the k-space and real-space boxes
+            fftwf_complex *HIRES_box = (fftwf_complex *) fftwf_malloc(sizeof(fftwf_complex)*KSPACE_NUM_PIXELS);
+            fftwf_complex *LOWRES_box = (fftwf_complex *) fftwf_malloc(sizeof(fftwf_complex)*HII_KSPACE_NUM_PIXELS);
+
+#pragma omp parallel shared(HIRES_box) private(i,j,k) num_threads(user_params->N_THREADS)
+            {
+#pragma omp for
+                for (i=0; i<user_params->DIM; i++){
+                    for (j=0; j<user_params->DIM; j++){
+                        for (k=0; k<=MIDDLE; k++){
+                            HIRES_box[C_INDEX(i, j, k)] = 0.0;
+                        }
+                    }
+                }
+            }
+
+            sprintf(wisdom_filename,"real_to_complex_DIM%d_NTHREADS%d.fftwf_wisdom",user_params->DIM,user_params->N_THREADS);
+            if(fftwf_import_wisdom_from_filename(wisdom_filename)==0) {
+                // Going to need to construct an FFTW_Wisdom for this box
+                plan = fftwf_plan_dft_r2c_3d(user_params->DIM, user_params->DIM, user_params->DIM,
+                                             (float *)HIRES_box, (fftwf_complex *)HIRES_box, FFTW_PATIENT);
+                fftwf_execute(plan);
+
+                // Store the wisdom for later use
+                fftwf_export_wisdom_to_filename(wisdom_filename);
+            }
+            fftwf_destroy_plan(plan);
+
+            sprintf(wisdom_filename,"complex_to_real_DIM%d_NTHREADS%d.fftwf_wisdom",user_params->DIM,user_params->N_THREADS);
+            if(fftwf_import_wisdom_from_filename(wisdom_filename)==0) {
+
[…]
