### 2020-07-23 1ecc6bf
Add error code for a missing file
