+++ src/py21cmfast/src/GenerateICs.c
@@ -289,20 +289,8 @@ int ComputeInitialConditions(
             fftwf_execute(plan);
         }
         else {
-            plan = fftwf_plan_dft_c2r_3d(user_params->DIM, user_params->DIM, user_params->DIM,
-                                         (fftwf_complex *)HIRES_box, (float *)HIRES_box, FFTW_PATIENT);
-            fftwf_execute(plan);
-
-            // Store the wisdom for later use
-            fftwf_export_wisdom_to_filename(wisdom_filename);
-
-            // copy over unfiltered box
-            memcpy(HIRES_box, HIRES_box_saved, sizeof(fftwf_complex)*KSPACE_NUM_PIXELS);
-
-            fftwf_destroy_plan(plan);
-            plan = fftwf_plan_dft_c2r_3d(user_params->DIM, user_params->DIM, user_params->DIM,
-                                         (fftwf_complex *)HIRES_box, (float *)HIRES_box, FFTW_WISDOM_ONLY);
-            fftwf_execute(plan);
+            LOG_ERROR("Cannot locate FFTW Wisdom: %s file not found",wisdom_filename);
+            Throw(FileError);
         }
     }
     else {
@@ -417,32 +405,12 @@ int ComputeInitialConditions(
                 if(fftwf_import_wisdom_from_filename(wisdom_filename)!=0) {
                     plan = fftwf_plan_dft_r2c_3d(user_params->DIM, user_params->DIM, user_params->DIM,
                                                  (float *)HIRES_box_vcb_x, (fftwf_complex *)HIRES_box_vcb_x, FFTW_WISDOM_ONLY);
+                    fftwf_execute(plan);
                 }
                 else {
-                    // Going to need to construct an FFTW_Wisdom for this box. Now its time to allocate the memory to save a copy of the box which gets
-                    // destroyed on FFTW Wisdom creation
-                    HIRES_box_vcb_saved = (fftwf_complex *) fftwf_malloc(sizeof(fftwf_complex)*KSPACE_NUM_PIXELS);
-
-                    // Make a copy of the x-direction velocity
-                    memcpy(HIRES_box_vcb_saved, HIRES_box_vcb_x, sizeof(fftwf_complex)*KSPACE_NUM_PIXELS);
-
-                    plan = fftwf_plan_dft_r2c_3d(user_params->DIM, user_params->DIM, user_params->DIM,
-                                             (float *)HIRES_box_vcb_x, (fftwf_complex *)HIRES_box_vcb_x, FFTW_PATIENT);
-                    fftwf_execute(plan);
-
-                    // Store the wisdom for later use
-                    fftwf_export_wisdom_to_filename(wisdom_filename);
-
-                    // return copy of the x-direction velocity that we saved
-                    memcpy(HIRES_box_vcb_x, HIRES_box_vcb_saved, sizeof(fftwf_complex)*KSPACE_NUM_PIXELS);
-
-                    plan = fftwf_plan_dft_r2c_3d(user_params->DIM, user_params->DIM, user_params->DIM,
-                                                 (float *)HIRES_box_vcb_x, (fftwf_complex *)HIRES_box_vcb_x, FFTW_WISDOM_ONLY);
-
-                    // No longer need the saved box, we only enter here once
-                    fftwf_free(HIRES_box_vcb_saved);
+                    LOG_ERROR("Cannot locate FFTW Wisdom: %s file not found",wisdom_filename);
+                    Throw(FileError);
                 }
-                fftwf_execute(plan);
 
                 plan = fftwf_plan_dft_r2c_3d(user_params->DIM, user_params->DIM, user_params->DIM,
[…]
