### 2020-07-23 20179d2
Add create FFTW wisdom function to all necessary functions
