### 2020-07-23 8d50acb
Remove FFTW creation if file doesn't exist (IoniseBox)
