### 2020-07-23 68ea4fb
Fix usage of destroy when no plan created
