### 2020-07-24 #2502 @steven-murray
Thanks @dfm! It looks like one of the figures is not rendering, and a couple of the references. I'll fix them up ASAP.
