### 2020-07-28 a405d25
paper: fix image ref and citations
