### 2020-07-28 36a97ca
paper: fix image ref and citations
