### 2020-07-28 cd26842
Make minor modifications (changelog + minor bug fixes)
