+++ src/py21cmfast/outputs.py
@@ -14,6 +14,7 @@ parameter objects necessary to define it.
 import os
 import warnings
 from hashlib import md5
+from typing import Sequence
 
 import h5py
 import numpy as np
@@ -634,6 +635,7 @@ class Coeval(_HighLevelOutput):
         ionized_box: IonizedBox,
         brightness_temp: BrightnessTemp,
         ts_box: [TsBox, None] = None,
+        cache_files: [dict, None] = None,
         photon_nonconservation_data=None,
         _globals=None,
     ):
@@ -653,6 +655,8 @@ class Coeval(_HighLevelOutput):
         self.brightness_temp_struct = brightness_temp
         self.spin_temp_struct = ts_box
 
+        self.cache_files = cache_files
+
         self.photon_nonconservation_data = photon_nonconservation_data
         # A *copy* of the current global parameters.
         self.global_params = _globals or dict(global_params.items())
@@ -669,6 +673,107 @@ class Coeval(_HighLevelOutput):
                 continue
             for field in box.fieldnames:
                 setattr(self, field, getattr(box, field))
+
+    def get_cached_data(
+        self, kind: str, redshift: float, load_data: bool = False
+    ) -> _OutputStruct:
+        """
+        Return an OutputStruct object which was cached in creating this Coeval box.
+
+        Parameters
+        ----------
+        kind
+            The kind of object: "init", "perturb", "spin_temp", "ionize" or "brightness"
+        redshift
+            The (approximate) redshift of the object to return.
+        load_data
+            Whether to actually read the field data of the object in (call ``obj.read()``
+            after this function to do this manually)
+
+        Returns
+        -------
+        output
+            The output struct object.
+        """
+        if self.cache_files is None:
+            raise AttributeError(
+                "No cache files were associated with this Coeval object."
+            )
+
+        files = self.cache_files.get(kind, {})
+        # files is a list of tuples of (redshift, filename)
+
+        redshifts = np.array([f[0] for f in files])
+
+        indx = np.argmin(np.abs(redshifts) - redshift)
+        fname = files[indx][1]
+
+        if not os.path.exists(fname):
+            raise IOError(
+                "The cached file you requested does not exist (maybe it was removed?)."
+            )
+
+        kinds = {
+            "init": InitialConditions,
+            "perturb": PerturbedField,
+            "ionize": IonizedBox,
+            "spin_temp": TsBox,
+            "brightness": BrightnessTemp,
+        }
+        cls = kinds[kind]
+
+        return cls.from_file(fname, load_data=load_data)
+
+    def gather(
+        self,
+        fname,
+        kinds: [Sequence, None] = None,
+        clean: [bool, dict] = False,
+        direc=None,
+    ):
+        """Gather the cached data associated with this object into its file."""
+        kinds = kinds or ["init", "perturb", "ionize", "spin_temp", "brightness"]
+
+        if clean and not hasattr(clean, "__len__"):
+            clean = kinds
+
+        if any(c not in kinds for c in clean):
+            raise ValueError(
+                "You are trying to clean cached items that you will not be gathering."
+            )
+
+        direc = os.path.expanduser(direc or config["direc"])
+
+        if fname is None:
+            fname = self.get_unique_filename()
+
+        if not os.path.isabs(fname):
+            fname = os.path.abspath(os.path.join(direc, fname))
+
+        for kind in kinds:
+            redshifts = (f[0] for f in self.cache_files[kind])
+            for i, z in enumerate(redshifts):
+                cache_fname = self.cache_files[kind][i][1]
+
+                obj = self.get_cached_data(kind, redshift=z, load_data=True)
+                with h5py.File(fname, "a") as fl:
+                    cache = (
[…]
