### 2020-09-08 #2582 @sambit-giri
Hi @dfm
I am unable to edit the reviewer's checklist. I don't remember if I had accepted the invitation. I apologise for it. When I try to do it now, I get the following message:
"Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account."
