### 2020-09-08 #2582 @sambit-giri
Hi @steven-murray
Here are a few comments on the software paper:

1. In paragraph 1, you say “Between these extremes lies an especially versatile middle-ground: fast semi-numerical models that approximate the full 3D evolution of the relevant fields: density, velocity, temperature, ionization, and radiation.” It is not clear which radiation you are talking about.
2. In paragraph 1, you say  “...so much so that they can be used to produce thousands of realizations on scales comparable to those observable by upcoming telescopes, in order....”. Can you mention the observables and few upcoming telescopes? Do you just mean the 21-cm signal and the upcoming radio telescopes? 21cmFast has also been used to put constraints based on other observations, such as quasar spectra and CMB observations. What about them?
3. The caption in Figure 1 does not describe the figure completely. The “component fields” are labelled with symbols in the figure. They have not been defined in the text and therefore are difficult to understand.
4. In paragraph 4, you say “This paper presents 21cmFAST v3+, which enables these essential new features. While keeping the same core functionality of previous versions of 21cmFAST, it…”. It is unclear which are the “new features” and which are the “same core functionality”.
5. There is an error in your citation in the fourth line of the list in paragraph 6.
6. In paragraph 7, I suggest you cite some example papers about milli-charged dark matter models and forward-modelled CMB auxiliary data.
7. Please cite the 21CMMC paper.
8. Please add a citation for “CLASS Boltzmann code”.
9. In Figure 3, please mention that ‘xH’ is the mass/volume average neutral fraction.
