### 2020-09-15 #2582 @qyx268
Hey @sambit-giri, thanks for reviewing this! I could answer the following points. @steven-murray Could you please add these in the paper?

> 3. The caption in Figure 1 does not describe the figure completely. The “component fields” are labelled with symbols in the figure. They have not been defined in the text and therefore are difficult to understand.

Lightcones for the reference model including, from left to right:
 			(i) overdensity ($\Delta_{\rm cell}$);
 			(ii) {\lya} ($J_{\alpha, {\rm eff}}{\times}\rm s^{-1}Hz^{-1}cm^{-2}sr^{-1}$);
 			(iii) LW ($J^{\rm 21}_{\rm LW,eff}{\times}{\rm 10^{{-}21}{\rm erg\ s^{-1}\ Hz^{-1}\ cm^{-2}\ sr^{-1}}}$);
 			(iv) X-ray heating ($\varepsilon_{X}$ in units of $k_\mathrm{B} {\rm K Gyr}^{-1}$);
 			(v) locally averaged UVB ($\bar{\Gamma}_{\rm ion}$ in units of $10^{-12}\rm{s}^{-1}$);
 			(vi) critical halo mass for star formation in ACGs ($M_{\rm crit}^{\rm atom}$/${\msol}$);
 			(vii) critical halo mass for star formation in MCGs ($M_{\rm crit}^{\rm mol}$/${\msol}$);
 			(viii) cumulative number of recombinations per baryon ($n_{\rm rec}$);
 			(ix) neutral hydrogen fraction ($x_{\hone}$); and
 			(x) the 21-cm brightness temperature ($\delta T_{\rm b}$ in units of ${\rm mK}$).\newline
 			{(A high-resolution version of this figure is available at \url{http://homepage.sns.it/mesinger/Media/light-cones_minihalo.png}.})

> 5. There is an error in your citation in the fourth line of the list in paragraph 6.

```
@ARTICLE{2020MNRAS.495..123Q,
       author = {{Qin}, Yuxiang and {Mesinger}, Andrei and {Park}, Jaehong and
         {Greig}, Bradley and {Mu{\~n}oz}, Julian B.},
        title = "{A tale of two sites - I. Inferring the properties of minihalo-hosted galaxies from current observations}",
      journal = {\mnras},
     keywords = {galaxies: high-redshift, intergalactic medium, dark ages, reionization, first stars, diffuse radiation, early Universe, cosmology: theory, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
         year = 2020,
        month = apr,
       volume = {495},
       number = {1},
        pages = {123-140},
          doi = {10.1093/mnras/staa1131},
archivePrefix = {arXiv},
       eprint = {2003.04442},
 primaryClass = {astro-ph.CO},
       adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.495..123Q},
      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
}
```
> 8. Please add a citation for “CLASS Boltzmann code”.

```
@ARTICLE{Lesgourgues2011arXiv1104.2932L,
	author = {{Lesgourgues}, Julien},
	title = "{The Cosmic Linear Anisotropy Solving System (CLASS) I: Overview}",
	journal = {arXiv e-prints},
	keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Extragalactic Astrophysics},
	year = 2011,
	month = apr,
	eid = {arXiv:1104.2932},
	pages = {arXiv:1104.2932},
	archivePrefix = {arXiv},
	eprint = {1104.2932},
	primaryClass = {astro-ph.IM},
	adsurl = {https://ui.adsabs.harvard.edu/abs/2011arXiv1104.2932L},
	adsnote = {Provided by the SAO/NASA Astrophysics Data System}
}
```
