### 2020-09-22 371e1d8
doc: address Sambit's review comments
