+++ joss-paper/paper.bib
@@ -119,3 +119,92 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2019PhRvD.100f3538M},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{qin2020,
+       author = {{Qin}, Yuxiang and {Mesinger}, Andrei and {Park}, Jaehong and
+         {Greig}, Bradley and {Mu{\~n}oz}, Julian B.},
+        title = "{A tale of two sites - I. Inferring the properties of minihalo-hosted galaxies from current observations}",
+      journal = {\mnras},
+     keywords = {galaxies: high-redshift, intergalactic medium, dark ages, reionization, first stars, diffuse radiation, early Universe, cosmology: theory, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2020,
+        month = apr,
+       volume = {495},
+       number = {1},
+        pages = {123-140},
+          doi = {10.1093/mnras/staa1131},
+archivePrefix = {arXiv},
+       eprint = {2003.04442},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.495..123Q},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Lesgourgues2011,
+	author = {{Lesgourgues}, Julien},
+	title = "{The Cosmic Linear Anisotropy Solving System (CLASS) I: Overview}",
+	journal = {arXiv e-prints},
+	keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Extragalactic Astrophysics},
+	year = 2011,
+	month = apr,
+	eid = {arXiv:1104.2932},
+	pages = {arXiv:1104.2932},
+	archivePrefix = {arXiv},
+	eprint = {1104.2932},
+	primaryClass = {astro-ph.IM},
+	adsurl = {https://ui.adsabs.harvard.edu/abs/2011arXiv1104.2932L},
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Munoz2018,
+       author = {{Mu{\~n}oz}, Julian B. and {Dvorkin}, Cora and {Loeb}, Abraham},
+        title = "{21-cm Fluctuations from Charged Dark Matter}",
+      journal = {\prl},
+     keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, High Energy Physics - Phenomenology},
+         year = 2018,
+        month = sep,
+       volume = {121},
+       number = {12},
+          eid = {121301},
+        pages = {121301},
+          doi = {10.1103/PhysRevLett.121.121301},
+archivePrefix = {arXiv},
+       eprint = {1804.01092},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018PhRvL.121l1301M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{qin2020a,
+       author = {{Qin}, Yuxiang and {Poulin}, Vivian and {Mesinger}, Andrei and
+         {Greig}, Bradley and {Murray}, Steven and {Park}, Jaehong},
+        title = "{Reionization inference from the CMB optical depth and E-mode polarization power spectra}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2020,
[…]
