### 2020-09-23 #2582 @steven-murray
@sambit-giri: I've updated the paper trying to address your points (see https://github.com/21cmfast/21cmFAST/pull/105/commits/371e1d8d502cec94408e4c96a423f4e1930ecf40 for the diff).

In brief:

> 1. In paragraph 1, you say “Between these extremes lies an especially versatile middle-ground: fast semi-numerical models that approximate the full 3D evolution of the relevant fields: density, velocity, temperature, ionization, and radiation.” It is not clear which radiation you are talking about.

I've added that it is Lyman-alpha and 21-cm radiation.

> 2. In paragraph 1, you say “...so much so that they can be used to produce thousands of realizations on scales comparable to those observable by upcoming telescopes, in order....”. Can you mention the observables and few upcoming telescopes? Do you just mean the 21-cm signal and the upcoming radio telescopes? 21cmFast has also been used to put constraints based on other observations, such as quasar spectra and CMB observations. What about them?

I've restricted it to low-frequency radio telescopes. While other observables are also useful in constraining astrophysics using 21cmFAST, I wanted to focus on the primary motivation.

> 3. The caption in Figure 1 does not describe the figure completely. The “component fields” are labelled with symbols in the figure. They have not been defined in the text and therefore are difficult to understand.

Thanks for picking this up, I've added captions to all the figures, and this one describes the components.

> 4. In paragraph 4, you say “This paper presents 21cmFAST v3+, which enables these essential new features. While keeping the same core functionality of previous versions of 21cmFAST, it…”. It is unclear which are the “new features” and which are the “same core functionality”.

I've changed the wording here to be the "essential guiding principals", and it refers back to the end of the previous paragraph. It then goes on to describe the new features.

> 5. There is an error in your citation in the fourth line of the list in paragraph 6.

Fixed, thanks!

> 6. In paragraph 7, I suggest you cite some example papers about milli-charged dark matter models and forward-modelled CMB auxiliary data.

Added, thanks!

> 7. Please cite the 21CMMC paper.
> 8. Please add a citation for “CLASS Boltzmann code”.
> 9. In Figure 3, please mention that ‘xH’ is the mass/volume average neutral fraction.

All done.
