### 2020-09-23 e5e47c9
doc: add Andrei's acknolwedgement
