+++ .isort.cfg
@@ -1,10 +1,8 @@
 [settings]
-line_length=88
-indent='    '
 skip=.tox,.venv,build,dist
 known_standard_library=setuptools,pkg_resources
 known_test=pytest
 known_first_party=py21cmfast
 sections=FUTURE,STDLIB,COMPAT,TEST,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
 default_section=THIRDPARTY
-multi_line_output=3
+profile=black
