### 2020-09-23 d103488
doc: add details about installing via conda
