### 2020-09-24 8016e98
doc: omit Jpritch from acknowledgments
