+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+dev
+------
+Added
+~~~~~
+* ``coeval_callback`` fflag to the ``run_lightcone``. Gives the ability to run arbitrary code
+  on each ``Coeval`` box for each node redshift.
+
 v3.0.2
 ------
 
