### 2020-09-26 fa9eb2b
write test for coeval_callback
