### 2020-09-26 #2582 @sambit-giri
Hi @steven-murray

Can you make the GitHub repository self-sustained? You can add a short summary of what 21cmFast can do. It will be helpful for new users. You should also add some of the easiest installation method (for more details you provide a link to your readthedocs installation page), instructions about automated tests post-installation and community guidelines (contribution, raising issue etc).

I installed the package in MacOSX. The instructions in the readthedocs page was not enough. I had to refer to the GitHub issue you mention. Is it possible to improve the instructions by summarising a few of the important information discussed in the issue. Also, "conda install -c conda-forge 21cmFAST" worked better for me. Please mention this method in the readthedocs page and GitHub readme.
