### 2020-09-26 #2582 @sambit-giri
@steven-murray

My automated tests failed. I am attaching a file containing the output I received while running `python -m pytest`.

[pytest_output.txt](https://github.com/openjournals/joss-reviews/files/5286637/pytest_output.txt)
