### 2020-09-26 #2582 @sambit-giri
@steven-murray
Thanks for updating the software paper.

Just a final comment on that. The link given for Figure 1 is wrong. There is a hyphen in `light-cones` in the link given, but I see the high-resolution image at http://homepage.sns.it/mesinger/Media/lightcones_minihalo.png
