### 2020-09-29 049cb88
exception handling for coeval_callback
