### 2020-09-29 f9ec8fe
Add basic description to readme
