### 2020-09-29 a6dfd9d
doc: add installation instructions
