+++ README.rst
@@ -19,8 +19,9 @@
 
 **A semi-numerical cosmological simulation code for the radio 21-cm signal.**
 
-.. image:: joss-paper/yuxiangs-small-plots.png
-    :target:
+.. image:: joss-paper/yuxiangs-plot-small.png
+    :target: http://homepage.sns.it/mesinger/Media/lightcones_minihalo.png
+
 
 This is the official repository for `21cmFAST`: a semi-numerical code that is able to
 produce 3D cosmological realisations of many physical fields in the early Universe.
@@ -52,10 +53,22 @@ New Features in 3.0.0+
 * Comprehensive test suite.
 * Strict `semantic versioning <https://semver.org>`_.
 
+Installation
+============
+We support Linux and MacOS (please let us know if you are successful in installing on
+Windows!). On these systems, the simplest way to get `21cmFAST` is by using
+`<https://www.anaconda.com/>conda`_::
+
+    conda install -c conda-forge 21cmFAST
+
+`21cmFAST` is also available on PyPI, so that `pip install 21cmFAST` also works. However,
+it depends on some external (non-python) libraries that may not be present, and so this
+method is discouraged unless absolutely necessary. If using `pip` to install `21cmFAST`
+(especially on MacOS), we thoroughly recommend reading the detailed
+`<https://21cmfast.readthedocs.io/en/latest/installation.html>installation instructions`_.
 
 Documentation
 =============
-
 Full documentation (with examples, installation instructions and full API reference)
 found at https://21cmfast.readthedocs.org.
 
