### 2020-09-29 6e9fd64
doc: fix link to yuxiang's plot
