### 2020-09-29 #2582 @steven-murray
Hi @sambit-giri, thanks for the very useful suggestions!

> Can you make the GitHub repository self-sustained? You can add a short summary of what 21cmFast can do. It will be helpful for new users.

I've made these changes in https://github.com/21cmfast/21cmFAST/pull/172 (hopefully merged soon). The description is short but hopefully helpful.

> You should also add some of the easiest installation method (for more details you provide a link to your readthedocs installation page),

Done in the above PR as well.

> instructions about automated tests post-installation and community guidelines (contribution, raising issue etc).

Running automated tests post-installation is only possible if one clones the repo itself and installs that way. We expect that to happen only for those who want to be developers, in which case the developer documentation outlines how to run the tests.
Also, we already have community guidelines in the organization's `.github` repo, which should show up when a new user tries to create an issue or PR. Agreed that it is less easy to find those instructions manually this way, but it means we have the same instructions for 21cmMC without redundant files.

> I installed the package in MacOSX. The instructions in the readthedocs page was not enough. I had to refer to the GitHub issue you mention. Is it possible to improve the instructions by summarising a few of the important information discussed in the issue. Also, conda install -c conda-forge 21cmFAST worked better for me. Please mention this method in the readthedocs page and GitHub readme.

The installation page on RTD has been updated to include the simpler conda install (however, I've just noticed that the RTD build is failing, hence you didn't see it -- will fix that ASAP). Installing on MacOS without `conda` is just hard, and it would be difficult to summarise the different cases succinctly. I feel like referencing the issue is the best we can do for now.

Concerning your failed tests -- can you inform me of which version you're testing (master branch or otherwise?). CI is passing all tests on master currently (on MacOS as well), so it is strange that they fail for you.

I've also updated the link in the paper, thanks!
