### 2020-09-29 932a54a
fix: only raise error on non-available table-path when using it
