+++ src/py21cmfast/inputs.py
@@ -16,6 +16,7 @@ import contextlib
 import logging
 from astropy.cosmology import Planck15
 from os import path
+from pathlib import Path
 
 from ._utils import StructInstanceWrapper, StructWithDefaults
 from .c_21cmfast import ffi, lib
@@ -288,14 +289,19 @@ class GlobalParams(StructInstanceWrapper):
     def __init__(self, wrapped, ffi):
         super().__init__(wrapped, ffi)
 
-        table_path = path.join(path.expanduser("~"), ".21cmfast")
-        EXTERNALTABLES = ffi.new("char[]", table_path.encode())
-        self.external_table_path = EXTERNALTABLES
+        self._table_path = Path.home() / ".21cmfast"
+        EXTERNALTABLES = ffi.new("char[]", str(self._table_path).encode())
+        self._external_table_path = EXTERNALTABLES
 
-        if not path.exists(table_path):
+    @property
+    def external_table_path(self):
+        """An ffi char pointer to the path to which external tables are kept."""
+        if not self._table_path.exists():
             raise IOError(
-                f"Found no user data directory for 21cmFAST! Should be at {table_path}"
+                f"Found no user data directory for 21cmFAST! Should be at {self._table_path}."
+                f"Try re-installing 21cmFAST. "
             )
+        return self._external_table_path
 
     @contextlib.contextmanager
     def use(self, **kwargs):
@@ -313,7 +319,7 @@ class GlobalParams(StructInstanceWrapper):
         this_attr_upper = {k.upper(): k for k in self.keys()}
 
         for k, val in kwargs.items():
-            if not k.upper() in this_attr_upper:
+            if k.upper() not in this_attr_upper:
                 raise ValueError(
                     "{} is not a valid parameter of global_params".format(k)
                 )
