### 2020-09-29 9919ae2
fix: add setter for external_table_pat
