### 2020-09-30 a7b5172
exception and warning handling for coeval_callback
