### 2020-09-30 2529413
test for coeval_callback exception
