+++ tests/test_wrapper.py
@@ -346,3 +346,24 @@ def test_coeval_callback_redshifts(ic, redshift, max_redshift, perturb_field):
         for i in coeval_callback_redshifts[1:]
     ]
     assert np.allclose(coeval_output, computed_redshifts)
+
+
+def test_coeval_callback_exceptions(ic, redshift, max_redshift, perturb_field):
+    # should output warning in logs and not raise an error
+    lc, coeval_output = wrapper.run_lightcone(
+        init_box=ic,
+        perturb=perturb_field,
+        max_redshift=max_redshift,
+        coeval_callback=lambda x: 1 / int(x.redshift - redshift),
+        coeval_callback_redshifts=[max_redshift, redshift],
+    )
+    # should raise an error
+    with pytest.raises(RuntimeError) as excinfo:
+        lc, coeval_output = wrapper.run_lightcone(
+            init_box=ic,
+            perturb=perturb_field,
+            max_redshift=max_redshift,
+            coeval_callback=lambda x: 1 / 0,
+            coeval_callback_redshifts=[max_redshift, redshift],
+        )
+    assert "coeval_callback computation failed on first trial" in str(excinfo.value)
