### 2020-09-30 c72e88d
change test function
