### 2020-10-01 #2582 @sultanier
Hi @steven-murray, seems like a very nice package, congrats!

 - The installation went great as described, however just small suggestion to add might help new users
to computing clusters that these dependencies like gsl, fftw..etc might already be installed so you might
add: use e.g. module load gsl ....

- I like the CLI interactive mode with the 21cmfast command but I was unable to run it.
Could you give a complete example of how to use it? Also would it be possible to add the whole parameter list here and their meanings? because the config file isn't complete ~/.21cmfast/runconfig_example.yml, and should one copies the config file to another location  or has to be there ..etc? Would be nice to include an example of a complete config file with all parameters and their definitions so a new user would know how to specify them, and where to store the generated simulation output.  I see in the following page you have tutorial of how to use the package within python, but some users, like me, might prefer the interactive mode with the 21cmfast command.
