### 2020-10-01 #2582 @sultanier
@steven-murray, how do we see all options for the 'kind' argument in the coeval_sliceplot, I couldn't find this in the API. the actual spelling might be different, also they aren't in the flag_options: coeval8.flag_options:
FlagOptions(INHOMO_RECO:False, M_MIN_in_Mass:False, PHOTON_CONS:False, SUBCELL_RSD:False, USE_HALO_FIELD:False, USE_MASS_DEPENDENT_ZETA:False, USE_MINI_HALOS:False, USE_TS_FLUCT:False)
