### 2020-10-01 #2582 @sultanier
@steven-murray the paper reads very nice, I think it misses how fast is this package compared to just running 21cmFAST directly? how the linking between python and C is achieved (OS module, Ctypes, Cython...etc)? The community would be interested to know the time spent for few examples: e.g. to to run a lightcone from z=20 to z=6 for box size 100 Mpc the code spends ...etc this is also relevant for users to have a rough estimate if they want to run MCMC with that. Otherwise looking good congrats again.
