### 2020-10-02 c68eaed
doc: add significant usage docs to readme
