### 2020-10-02 c97d2b9
doc: improve the runconfig example and basic usage notebook
