+++ user_data/runconfig_example.yml
@@ -1,8 +1,17 @@
-# Specify parameters that 21cmFAST uses. These are broken into user, flags, astro and
-# cosmo parameters. Not all options are specified here.
+# Specify parameters that 21cmFAST uses in the CLI.
+# These are broken into user, flags, astro and cosmo parameters.
+# Note that not all options are specified here. A full list of parameters for each of
+# the dictionaries can be found here: https://21cmfast.readthedocs.io/en/latest/reference/_autosummary/py21cmfast.inputs.html
+
 user_params:
-    BOX_LEN: 150.0
-    HII_DIM: 50
+  BOX_LEN: 150.0
+  HII_DIM: 50
 
 flag_options:
   INHOMO_RECO: false
+
+cosmo_params:
+  SIGMA_8: 0.8
+
+astro_params:
+  HII_EFF_FACTOR: 30.0
