### 2020-10-02 0ea06e3
feat: add get_fields classmethod to Coeval to be able to obtain a list of kinds
