### 2020-10-02 07edb8d
doc: better suggested box size in simple usage example
