### 2020-10-02 #2582 @steven-murray
Thanks @sultanier for the very useful comments on the code! I've tried to address most of your comments in this PR: https://github.com/21cmfast/21cmFAST/pull/174

> The installation went great as described, however just small suggestion to add might help new users
   to computing clusters that these dependencies like gsl, fftw..etc might already be installed so you might
   add: use e.g. module load gsl ....

Thanks, I added a `HPC` section to the installation docs.

> I like the CLI interactive mode with the 21cmfast command but I was unable to run it.
> Could you give a complete example of how to use it? Also would it be possible to add the whole parameter list here and their        meanings? because the config file isn't complete ~/.21cmfast/runconfig_example.yml, and should one copies the config file to another location or has to be there ..etc? Would be nice to include an example of a complete config file with all parameters and their definitions so a new user would know how to specify them, and where to store the generated simulation output. I see in the following page you have tutorial of how to use the package within python, but some users, like me, might prefer the interactive mode with the 21cmfast command.

Firstly, could you let me know what went wrong for you with the CLI command?

I've updated the documentation in the README to have a quick guide on using the CLI. It references the user to https://21cmfast.readthedocs.io/en/latest/reference/_autosummary/py21cmfast.inputs.html to get a list of all the possible inputs and their documentation. I've also put this URL into the `runconfig_example.yml`. The config file location is specified with `--config=FILE` on the CLI, which is also now documented in the README. I would argue against including a COMPLETE config, as it would be doubling-up the documentation of the code itself (and may get out of date). As for storing the generated output, the README now has some description of this, and you are able to save the (coeval or lightcone) file explicitly. I hope these things make it easier to use!

> could you add in the beginning of section: Running and Plotting Coeval Cubes after "Clear the cache so that we get the same results for the notebook every time (don’t worry about this for now Also, set the default output directory to _cache/:" create dir, mkdir '_cache' where you want to run this jupyternotebook, otherwise it says _cache not found.

Done, thanks!

 > when you save files, is there anyway to make the saved files names more meaningful, for instance in this saved box name "BrightnessTemp_8da85e6e1d0e198f6a7f0f6b3af0a9a2_r12345.h5", I recognize that the last part in the files names is the seed number, is there any way for the user to specify the name? or the name can be changed to reflect other cosmo params?

The short answer is no, the automatically-saved cache files have a particular naming structure that is based on a hash of all the parameters that went into the simulation. We have tried to keep the most useful information in the filename (i.e. the kind of box it is, and the seed). Naming the file based on some subset of its parameters is not scalable, and would not do well as a cache. Of course, you can always copy whatever file you want to somewhere else on your filesystem and call it whatever you want! Also, we have made the `21cmfast query` tool, which lists each box and all the parameters that went into it, to make it a bit easier to know what each simulation really is.

Finally, the higher level outputs (Coeval and Lightcone) *do* have a `.save()` method which allows you to save the box with any filename you want. They are not part of the caching system, so this is OK.

> @steven-murray, how do we see all options for the 'kind' argument in the coeval_sliceplot, I couldn't find this in the API. the actual spelling might be different, also they aren't in the flag_options: coeval8.flag_options:
FlagOptions(INHOMO_RECO:False, M_MIN_in_Mass:False, PHOTON_CONS:False, SUBCELL_RSD:False, USE_HALO_FIELD:False, USE_MASS_DEPENDENT_ZETA:False, USE_MINI_HALOS:False, USE_TS_FLUCT:False)

This is actually an open issue (https://github.com/21cmfast/21cmFAST/issues/81). I've tried to partially address it in the PR I linked above. There is now a method that can be called on the `Coeval` class (you can do it directly, before having run a simulation): `Coeval.get_fields()`, that will list these kinds for you. I've also added this information to the `coeval_sliceplot` function docstring. It's not a full solution, as the output fields are not actually documented online, but I will leave that issue open to remind us to do this in the future.

> @steven-murray the paper reads very nice, I think it misses how fast is this package compared to just running 21cmFAST directly? how the linking between python and C is achieved (OS module, Ctypes, Cython...etc)? The community would be interested to know the time spent for few examples: e.g. to to run a lightcone from z=20 to z=6 for box size 100 Mpc the code spends ...etc this is also relevant for users to have a rough estimate if they want to run MCMC with that. Otherwise looking good congrats again.

This is a great suggestion. I will add another small section into the paper to describe this.
