### 2020-10-02 78ca1d5
doc: add performance notes to joss paper
