### 2020-10-05 4257bcb
doc: make it more clear that FFTW has to be the right version
