### 2020-10-12 a01cbba
Add in error that redshifts/xHI must be ascending for computing tau
