### 2020-10-12 #2582 @sultanier
thanks @steven-murray

for the CLI command, I was't sure how to run it because there was no detailed config file and the required parameter list was not mentioned. I will try again once you add a detailed config file. Let me know once you finish adding/implementing these suggestions/comments, so I can double check them for the last time before finishing up my checklist. thanks.
