### 2020-10-12 #2582 @steven-murray
@sultanier all the comments have been address in the current master (there are some formatting issues with the hyperlinks in the readme which are being addressed in a PR, but these shouldn't hamper your checks). The performance information to go into the paper itself is still a work in progress.
