### 2020-10-14 #2582 @sultanier
thanks @steven-murray,

I have finished checking all comments and everything looking good and working nicely! congrats!

One small fix is that these hyperlinks below for the config_example and inputs don't work (clicking on them doesn't directly take you to the files), but one can go to the repo and see these files. Just make sure these hyperlinks do actually work (if meant). I have completed all of my checklist but I am very curious to know the performance and speed for different examples, once you finish please generate the new pdf and let me know. Package is very beautifully made, congrats again!.

# hyperlinks.....
There is an example configuration file `here<user_data/runconfig_example.yml>`_ that you can build from. All input parameters are `documented here<https://21cmfast.readthedocs.io/en/latest/reference/_autosummary/py21cmfast.inputs.html>`_.
