### 2020-10-14 #2582 @steven-murray
Thanks @sultanier! There is a PR currently to fix those hyperlinks (should be merged in the next couple of days). We'll have those performance metrics into the paper very soon as well.
