### 2020-10-14 72b7917
doc: fill in performance notes
