### 2020-10-16 c6f1c2c
Make it more Pythonic
