### 2020-10-16 9b26588
doc: fixes to benchmarking notes
