### 2020-10-16 #2582 @steven-murray
@sultanier the updates to the hyperlinks have been merged. The benchmarks have also been added to the paper, but I can't seem to get it to compile. @dfm do you have any idea what could be going wrong there? The paper is definitely still there in the `joss-paper` branch...

@sambit-giri I'm not sure if all your questions/comments were addressed adequately?
