### 2020-10-16 8a811f6
doc update the benchmarking tables
