### 2020-10-16 7ae2bd9
docs: fix typos
