### 2020-10-16 #2582 @dfm
Pinging @openjournals/dev to see if there are any thoughts about why @whedon can't find the paper. I can confirm that [whedon.theoj.org](https://whedon.theoj.org/) is able to find it. Thanks!
