### 2020-10-16 #2582 @arfon
OK, I've figured out what's going on. I've not quite figured out the 'proper' fix here but I have a short-term solution. **tl;dr** – delete (or rename) the `joss-paper` folder in your default (`master`) branch and everything should be fine.

If you're interested, here's what's going on:

The issue here is that we're asking Whedon to act on the `joss-paper` branch but there's also a folder named `joss-paper` in the default (`master`) branch.

This confuses Git<sup>1</sup>, and I thought that [this commit](https://github.com/openjournals/whedon-api/commit/6c8d7d9127c6483e7c48c1097ff17d1876e77711) would address the issue (it does locally with my local Git version) but Whedon's Git is older and this flag doesn't seem to work there.

Ultimately I need to spend more time debugging this but I'm out of time this evening and don't want to block this further.

<sup>1</sup> https://stackoverflow.com/questions/25322335/git-change-branch-when-file-of-same-name-is-present
