### 2020-10-16 5c4e4ef
doc: some formatting fixes in the joss paper
