### 2020-10-16 65b666a
doc: make yuxiang's plot bigger in joss paper
