### 2020-10-19 #2582 @sambit-giri
@dfm I noticed that the author's list of the paper has been modified. I have worked with Catherine Watkinson in a project where both of us were primary authors (https://ui.adsabs.harvard.edu/abs/2019MNRAS.482.2653W/abstract). Does this violate the JOSS conflict of interest policy?
