### 2020-10-19 39a75e1
fix some major memory leak
