### 2020-10-19 98a6532
bug: freed r twice..
