### 2020-10-19 8f40eeb
should not free FgtrM_1DTable_linear (the same applys for other intepolation tables which are needed for all redshifts...)
