### 2020-10-19 c1261b4
Formatting tweaks to JOSS paper
