+++ joss-paper/paper.md
@@ -93,8 +93,8 @@ knowledge of the full system, or disrupting its workings.
 This lack of modularity within the code has led to
 widespread code "branching" as researchers hack new physical features of interest
 into the C code; the lack of a streamlined API has led derivative codes which run
-multiple realizations of `21cmFAST` simulations (such as the Monte Carlo simulator,
-`21CMMC` [@greig2015]) to re-write large portions of the code in order to serve their purpose.
+multiple realizations of `21cmFAST` simulations [such as the Monte Carlo simulator,
+`21CMMC` @greig2015] to re-write large portions of the code in order to serve their purpose.
 It is thus of critical importance, as the field moves forward in its understanding -- and
 the range and scale of physical models of interest continues to increase -- to
 reformulate the `21cmFAST` code in order to provide a fast, modular, well-documented,
@@ -141,7 +141,7 @@ The primary *new* usability features of `21cmFAST` v3+ are:
   remaining C global variables.
 * Comprehensive API documentation and tutorials.
 * Comprehensive test suite (and continuous integration).
-* Strict `semantic versioning <https://semver.org>`_.
+* Strict semantic versioning^[https://semver.org].
 
 While in v3 we have focused on the establishment of a stable and extendable infrastructure,
 we have also incorporated several new scientific features, appearing in separate papers:
