### 2020-10-19 44dac1a
Small typo in paper
