### 2020-10-19 #2582 @dfm
@steven-murray: I've opened a PR to the 21cmfast repo with a few minor formatting edits to the paper. Please take a look (and preferably accept) those changes and then go through the following steps:

1. I see now that there is also a `joss-paper-fixes` branch that will have one conflict with my PR. Please resolve those issues and confirm which branch the final submission will live on.
2. Get whedon to compile the paper and then read through the proofs carefully (especially checking things like author names and affiliations, but also general typos that we might have all missed) to make sure that you're happy. It's much harder to change things later! I believe that @sultan-hassan's username should be updated in the new version of the paper, but we should also check that.
3. Bump the version number and report the version number back here.
4. Create a tagged release and archive this release somewhere like Zenodo. The metadata of that archive (title and author list) **must exactly match the JOSS paper**. You can edit the metadata after creating the archive on most services. Report the DOI for the archive in this thread.

Let me know if you have any questions!
