### 2020-10-19 #2582 @dfm
@sambit-giri: I've checked with the editorial team and we've agreed that this does not constitute a conflict-of-interest in this case since you had already done most of your review and since it has been noted here. If @steven-murray or any of the co-authors have any concerns, please let me know here or offline, but otherwise we'll proceed. Thanks!
