### 2020-10-19 3aa8677
doc: fix formatting of DIM
