### 2020-10-19 #2582 @steven-murray
OK @dfm, I've merged the paper fixes, and the proof looks good.

As for the version update -- does that need to be a new version? Can we point to the last published version (3.0.2)? If not that's fine, but may be a day or so before we can release a new version.
