### 2020-10-19 a5046e6
bug: FreeTsInterpolationTables must take a cdata pointer
