+++ src/py21cmfast/wrapper.py
@@ -2367,7 +2367,7 @@ def run_coeval(
             photon_nonconservation_data = None
 
         if flag_options.USE_TS_FLUCT and user_params.USE_INTERPOLATION_TABLES:
-            lib.FreeTsInterpolationTables(flag_options)
+            lib.FreeTsInterpolationTables(flag_options())
 
         coevals = [
             Coeval(z, init_box, p, ib, _bt, st, photon_nonconservation_data)
