### 2020-10-19 #2582 @dfm
The version for the paper should represent the current state of the code base and docs. It looks to me like there have been a fair number of updates since 3.0.2 (July 2020) so I think you should make a new one. If you think otherwise, please let me know!
