### 2020-10-20 #2582 @steven-murray
You're right, @dfm. Made a new version.

New version is v3.0.3 and DOI is 10.5281/zenodo.4107189.
