### 2020-10-20 8721183
a few minor  improvement for the flags...
