+++ src/py21cmfast/src/IonisationBox.c
@@ -191,7 +191,7 @@ LOG_DEBUG("original redshift=%f, updated redshift=%f delta-z = %f", stored_redsh
     ArgBinWidth = (erfc_arg_max - erfc_arg_min)/((double)ERFC_NUM_POINTS - 1.);
     InvArgBinWidth = 1./ArgBinWidth;
 
-    if(INIT_ERFC_INTERPOLATION) {
+    if(!flag_options->USE_MASS_DEPENDENT_ZETA && INIT_ERFC_INTERPOLATION) {
 
         ERFC_VALS = calloc(ERFC_NUM_POINTS,sizeof(double));
         ERFC_VALS_DIFF = calloc(ERFC_NUM_POINTS,sizeof(double));
@@ -1690,13 +1690,11 @@ LOG_SUPER_DEBUG("freed fftw boxes");
 
     if (prev_redshift < 1){
         free(previous_ionize_box->z_re_box);
-        if (flag_options->USE_MASS_DEPENDENT_ZETA) {
-            if (flag_options->USE_MINI_HALOS){
-                free(previous_ionize_box->Gamma12_box);
-                free(previous_ionize_box->dNrec_box);
-                free(previous_ionize_box->Fcoll);
-                free(previous_ionize_box->Fcoll_MINI);
-            }
+        if (flag_options->USE_MASS_DEPENDENT_ZETA && flag_options->USE_MINI_HALOS){
+            free(previous_ionize_box->Gamma12_box);
+            free(previous_ionize_box->dNrec_box);
+            free(previous_ionize_box->Fcoll);
+            free(previous_ionize_box->Fcoll_MINI);
         }
     }
 
@@ -1706,7 +1704,7 @@ LOG_SUPER_DEBUG("freed fftw boxes");
         }
     }
 
-    if(INIT_ERFC_INTERPOLATION) {
+    if(!flag_options->USE_MASS_DEPENDENT_ZETA && INIT_ERFC_INTERPOLATION) {
 
         free(ERFC_VALS);
         free(ERFC_VALS_DIFF);
