### 2020-10-20 #2582 @dfm
Thanks! Can you edit the Zenodo meta data so that the title is "21cmFAST v3: A Python-integrated C code for generating 3D realizations of the cosmic 21cm signal"?
