### 2020-10-20 #2582 @dfm
@steven-murray: I see now that I missed that updated reference. It looks like https://ui.adsabs.harvard.edu/abs/2020MNRAS.499..550Q/abstract has now been published so can you update the bibtex entry to the published version?

EiC: I'll remove the `recommend-accept` label until @steven-murray fixes that. Sorry all about the false alarm!
