### 2020-10-20 055cbed
doc: update ref to Yuxiang's paper
