+++ joss-paper/paper.bib
@@ -121,23 +121,25 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{qin2020,
-       author = {{Qin}, Yuxiang and {Mesinger}, Andrei and {Park}, Jaehong and
-         {Greig}, Bradley and {Mu{\~n}oz}, Julian B.},
-        title = "{A tale of two sites - I. Inferring the properties of minihalo-hosted galaxies from current observations}",
+       author = {{Qin}, Yuxiang and {Poulin}, Vivian and {Mesinger}, Andrei and
+         {Greig}, Bradley and {Murray}, Steven and {Park}, Jaehong},
+        title = "{Reionization inference from the CMB optical depth and E-mode polarization power spectra}",
       journal = {\mnras},
-     keywords = {galaxies: high-redshift, intergalactic medium, dark ages, reionization, first stars, diffuse radiation, early Universe, cosmology: theory, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
+     keywords = {galaxies: high-redshift, intergalactic medium, cosmic background radiation, dark ages, reionization, first stars, early Universe, cosmology: theory, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2020,
-        month = apr,
-       volume = {495},
+        month = jul,
+       volume = {499},
        number = {1},
-        pages = {123-140},
-          doi = {10.1093/mnras/staa1131},
+        pages = {550-558},
+          doi = {10.1093/mnras/staa2797},
 archivePrefix = {arXiv},
-       eprint = {2003.04442},
+       eprint = {2006.16828},
  primaryClass = {astro-ph.CO},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.495..123Q},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.499..550Q},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
 
 @ARTICLE{Lesgourgues2011,
 	author = {{Lesgourgues}, Julien},
