### 2020-10-20 a4b027b
doc: update ref to Qin2020
