+++ joss-paper/paper.bib
@@ -120,8 +120,8 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{qin2020a,
-    author = {{Qin}, Yuxiang and {Mesinger}, Andrei and {Park}, Jaehong and
+@ARTICLE{qin2020,
+        author = {{Qin}, Yuxiang and {Mesinger}, Andrei and {Park}, Jaehong and
          {Greig}, Bradley and {Mu{\~n}oz}, Julian B.},
         title = "{A tale of two sites - I. Inferring the properties of minihalo-hosted galaxies from current observations}",
       journal = {\mnras},
@@ -174,7 +174,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{qin2020,
+@ARTICLE{qin2020a,
        author = {{Qin}, Yuxiang and {Poulin}, Vivian and {Mesinger}, Andrei and
          {Greig}, Bradley and {Murray}, Steven and {Park}, Jaehong},
         title = "{Reionization inference from the CMB optical depth and E-mode polarization power spectra}",
