### 2020-10-20 88bb67f
also freeSigmaMInterpTable
