### 2020-10-20 ced71d5
following the photoncon, only free when interpolation_tables_allocated
