+++ src/py21cmfast/src/21cmFAST.h
@@ -182,6 +182,7 @@ void free_TsCalcBoxes(struct UserParams *user_params, struct FlagOptions *flag_o
 void FreePhotonConsMemory();
 void FreeTsInterpolationTables(struct FlagOptions *flag_options);
 bool photon_cons_allocated = false;
+bool interpolation_tables_allocated = false;
 int SomethingThatCatches(bool sub_func);
 int FunctionThatCatches(bool sub_func, bool pass, double* result);
 void FunctionThatThrows();
