### 2020-10-20 58c3c8c
free density tables in the loop
