### 2020-10-20 424e0a0
doc: small whitespace change to paper.bib
