+++ joss-paper/paper.bib
@@ -121,7 +121,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{qin2020,
-        author = {{Qin}, Yuxiang and {Mesinger}, Andrei and {Park}, Jaehong and
+       author = {{Qin}, Yuxiang and {Mesinger}, Andrei and {Park}, Jaehong and
          {Greig}, Bradley and {Mu{\~n}oz}, Julian B.},
         title = "{A tale of two sites - I. Inferring the properties of minihalo-hosted galaxies from current observations}",
       journal = {\mnras},
