+++ src/py21cmfast/wrapper.py
@@ -2793,7 +2793,7 @@ def run_lightcone(
         else:
             photon_nonconservation_data = None
 
-        if flag_options.USE_TS_FLUCT and user_params.USE_INTERPOLATION_TABLES:
+        if flag_options.USE_TS_FLUCT and user_params.USE_INTERPOLATION_TABLES and lib.interpolation_tables_allocated:
             lib.FreeTsInterpolationTables(flag_options())
 
         out = (
