### 2020-10-21 #2582 @kyleniemeyer
Hi @steven-murray, I am the EIC on duty this week, doing some final checks before publishing.

I noticed that in the paper, journal names are not appearing for your references. Can you replace all the macros in your .bib file (e.g., `\apj`) with the spelled-out journal names?
